# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "backend",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/jobs": {
      "get": {
        "tags": [
          "candlescyther"
        ],
        "summary": "List all jobs.",
        "description": "Returns all jobs.",
        "operationId": "list_jobs",
        "responses": {
          "200": {
            "description": "List all jobs from jobs table.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/klines": {
      "get": {
        "tags": [
          "candlescyther"
        ],
        "summary": "List all klines per ticker.",
        "description": "Returns all klines for give ticker.",
        "operationId": "list_klines",
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List all klines for the ticker",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Kline"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ticker is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs": {
      "get": {
        "tags": [
          "candlescyther"
        ],
        "summary": "List all logs",
        "description": "Default 100 records.",
        "operationId": "list_logs",
        "responses": {
          "200": {
            "description": "List all logs from logs table.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEntry"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mf/sector": {
      "get": {
        "tags": [
          "candlescyther"
        ],
        "summary": "List moneyflow sector data.",
        "description": "Returns a 200 if the job is submitted.",
        "operationId": "list_mf_sector",
        "responses": {
          "200": {
            "description": "List moneyflow sector records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MoneyflowEastmoney"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Job runner error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "candlescyther"
        ],
        "summary": "Create moneyflow sector data.",
        "description": "Returns a 200 if the job is submitted.",
        "operationId": "create_mf_sector",
        "responses": {
          "200": {
            "description": "Job submitted"
          },
          "500": {
            "description": "Job runner error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "candlescyther"
        ],
        "summary": "Delete moneyflow sector data.",
        "description": "Returns a 200 if the job is submitted.",
        "operationId": "delete_mf_sector",
        "responses": {
          "200": {
            "description": "Delete moneyflow sector records successfully"
          },
          "500": {
            "description": "Job runner error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals": {
      "get": {
        "tags": [
          "candlescyther"
        ],
        "summary": "List all signals.",
        "description": "Returns all signals.",
        "operationId": "list_signals",
        "responses": {
          "200": {
            "description": "List all signals from signals table.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Signal"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/signals-us": {
      "get": {
        "tags": [
          "candlescyther"
        ],
        "summary": "List all signals for US stocks.",
        "description": "Returns all signals.",
        "operationId": "list_signals_us",
        "responses": {
          "200": {
            "description": "List all signals from signals table.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Signal"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stocks": {
      "get": {
        "tags": [
          "candlescyther"
        ],
        "summary": "List all stocks.",
        "description": "Returns all stocks.",
        "operationId": "list_stocks",
        "responses": {
          "200": {
            "description": "List all stocks from stocks table.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stock"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "candlescyther"
        ],
        "summary": "Create stocks with meta,klines,signals.",
        "description": "Returns a 200 if the job is submitted.",
        "operationId": "create_stocks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Job submitted"
          },
          "400": {
            "description": "Tickers is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Job runner error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "candlescyther"
        ],
        "summary": "Delete stock.",
        "operationId": "delete_stock",
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete stock and its records"
          },
          "400": {
            "description": "Ticker is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trigger/all": {
      "get": {
        "tags": [
          "candlescyther"
        ],
        "summary": "Trigger update of all.",
        "description": "Returns.",
        "operationId": "update_all",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trigger is init."
          },
          "400": {
            "description": "Missing code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "DatabaseError"
            ],
            "properties": {
              "DatabaseError": {
                "type": "string"
              }
            },
            "example": "failed to retrieve from database"
          },
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string"
              }
            },
            "example": "id = 1"
          },
          {
            "type": "object",
            "required": [
              "MissingInput"
            ],
            "properties": {
              "MissingInput": {
                "type": "string"
              }
            },
            "example": "require query param x"
          },
          {
            "type": "object",
            "required": [
              "Unauthorized"
            ],
            "properties": {
              "Unauthorized": {
                "type": "string"
              }
            },
            "example": "missing api key"
          },
          {
            "type": "object",
            "required": [
              "RunnerError"
            ],
            "properties": {
              "RunnerError": {
                "type": "string"
              }
            },
            "example": "job runner error"
          }
        ]
      },
      "CreateStockRequest": {
        "type": "object",
        "required": [
          "tickers"
        ],
        "properties": {
          "tickers": {
            "type": "string"
          }
        }
      },
      "Job": {
        "type": "object",
        "required": [
          "id",
          "job_status",
          "job_type",
          "payload",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "job_status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "payload": {},
          "updated_at": {
            "type": "string"
          }
        }
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "Pending",
          "Running",
          "Done",
          "Error"
        ]
      },
...
      "Kline": {
        "type": "object",
        "required": [
          "k_ticker",
          "k_date",
          "k_open",
          "k_high",
          "k_low",
          "k_close",
          "k_volume",
          "k_value"
        ],
        "properties": {
          "k_close": {
            "type": "number",
            "format": "double"
          },
          "k_date": {
            "type": "integer",
            "format": "int64"
          },
          "k_high": {
            "type": "number",
            "format": "double"
          },
          "k_low": {
            "type": "number",
            "format": "double"
          },
          "k_open": {
            "type": "number",
            "format": "double"
          },
          "k_ticker": {
            "type": "string"
          },
          "k_value": {
            "type": "number",
            "format": "double"
          },
          "k_volume": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "LogEntry": {
        "type": "object",
        "required": [
          "id",
          "log_timestamp",
          "log_level",
          "log_target",
          "log_message",
          "log_line"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "log_level": {
            "type": "integer",
            "format": "int64"
          },
          "log_line": {
            "type": "integer",
            "format": "int64"
          },
          "log_message": {
            "type": "string"
          },
          "log_target": {
            "type": "string"
          },
          "log_timestamp": {
            "type": "string"
          }
        }
      },
...
      "Signal": {
        "type": "object",
        "required": [
          "ticker",
          "kdj_k",
          "kdj_d",
          "boll_dist"
        ],
        "properties": {
          "boll_dist": {
            "type": "number",
            "format": "double"
          },
          "kdj_d": {
            "type": "number",
            "format": "double"
          },
          "kdj_k": {
            "type": "number",
            "format": "double"
          },
          "ticker": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}